check akey from header

+++ controllers/settings.js
@@ -5,6 +5,8 @@ const errors = require('../errors.json');
 const akeyMismatch = (settingsAKey, paramAKey) => settingsAKey != null && settingsAKey != paramAKey;
 
 const getSettings = asyncHandler(async(req, res, next) => {
+    if (req.params.akey != req.headers.AKey) return next(errors.AKEY_MISMATCH);
+
     res.json(await SettingsModel.findOneAndUpdate({
         akey: req.params.akey
     }, {
@@ -19,6 +21,7 @@ const getSettings = asyncHandler(async(req, res, next) => {
 const updateSettings = asyncHandler(async(req, res, next) => {
     const settingsObj = req.body || {};
 
+    if (req.params.akey != req.headers.AKey) return next(errors.AKEY_MISMATCH);
     if (akeyMismatch(settingsObj.akey, req.params.akey)) return next(errors.AKEY_MISMATCH);
 
     settingsObj.akey = req.params.akey;
@@ -33,6 +36,8 @@ const updateSettings = asyncHandler(async(req, res, next) => {
 });
 
 const deleteSettings = asyncHandler(async(req, res, next) => {
+    if (req.params.akey != req.headers.AKey) return next(errors.AKEY_MISMATCH);
+
     await SettingsModel.deleteOne({
         akey: req.params.akey
     });
